+++ pkg/sass_api/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 16.0.2
+
+* No user-visible changes.
+
 ## 16.0.1
 
 * No user-visible changes.